requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
firebase==3.0.1
firebase-admin==4.3.0
gcloud==0.17.0
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.1
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
idna==2.10
jws==0.1.3
msgpack==1.0.0
oauth2client==3.0.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
